Refactor segun nueva estrutura
