+++ android/app/src/main/res/layout/activity_view_tournaments.xml
@@ -12,7 +12,7 @@
         android:layout_height="wrap_content"
         android:id="@+id/TypeFilter"></Spinner>
     <ListView
-    android:id="@+id/sampleListView"
+    android:id="@+id/ListTournaments"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:layout_marginBottom="10dp"
